add model validation for quiz creation
